requirements.txt
# Este programa usa apenas bibliotecas padrão do Python
# Não são necessárias dependências externas
# tkinter - Interface gráfica (incluído no Python)
# pathlib - Manipulação de caminhos (incluído no Python)
# os, shutil, datetime - Operações de sistema (incluído no Python)
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Werkzeug==3.1.3
qrcode==7.4.2
Pillow==10.0.1
# Dependências para conversão de arquivos
pymupdf==1.23.8
python-docx==1.1.0
openpyxl==3.1.2
reportlab==4.0.7
# Dependência para download do YouTube
yt-dlp==2023.12.30
# Para variáveis de ambiente
python-dotenv==1.0.0
# Para CORS (se necessário)
Flask-CORS==4.0.0
# Para compressão
gunicorn==21.2.0
# Para conversão de áudio (opcional - instalar manualmente se necessário)
# ffmpeg-python==0.2.0
# Nota: pydub e moviepy foram removidos devido a problemas de compatibilidade com Python 3.12+
# Para conversão de áudio/vídeo, instale manualmente: pip install pydub moviepy 